pyproject.toml
[build-system]
requires = ["setuptools>=69", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "diff2tweet"
version = "0.1.0"
description = "Generate tweet drafts from git diffs and commit context."
readme = "README.md"
requires-python = ">=3.11"
license = {text = "MIT"}
authors = [
    {name = "bju12290"},
]
keywords = ["git", "twitter", "x", "llm", "openai", "anthropic", "developer-tools", "cli", "tweet"]
classifiers = [
    "Development Status :: 3 - Alpha",
    "Environment :: Console",
    "Intended Audience :: Developers",
    "License :: OSI Approved :: MIT License",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Programming Language :: Python :: 3.13",
    "Topic :: Software Development",
    "Topic :: Utilities",
]
dependencies = [
    "openai>=1,<2",
    "pydantic>=2,<3",
    "pydantic-settings>=2,<3",
    "PyYAML>=6,<7",
    "typer>=0.12,<1",
]

[project.scripts]
diff2tweet = "diff2tweet.cli:app"

[project.optional-dependencies]
anthropic = ["anthropic>=0.40,<1"]
gemini = ["google-genai>=1,<2"]
all-providers = ["anthropic>=0.40,<1", "google-genai>=1,<2"]
dev = [
    "pytest>=8,<9",
    "pytest-cov>=5,<6",
]

[project.urls]
Homepage = "https://github.com/bju12290/diff2tweet"
Repository = "https://github.com/bju12290/diff2tweet"
"Bug Tracker" = "https://github.com/bju12290/diff2tweet/issues"

[tool.setuptools]
package-dir = {"" = "src"}

[tool.setuptools.packages.find]
where = ["src"]

[tool.pytest.ini_options]
testpaths = ["tests"]
